requirements.txt
aiogram>=3.3,<4
python-dotenv>=1.0,<2
aiohttp>=3.12.15
asyncpg>=0.30.0
